Implementation of blogs and home button
The blogs are shown on the front page and clickable to the full blog, there is also a panel that recommends other blogs about the same subject. Also the Ace Tutoring logo becomes a home button when not on the home page - might separate this because it reduces the profile of the brand.