Implemented the student sign up page
I created a form for students or parents to express their interest in tutoring services. There are a number of validation steps including making sure at least one piece of contact information is included, that the subject selected is appropriate for the reported year level, and that the postcode is for an area reasonably close by if the customer wants in-person tutoring. Unit tests were built and were used to track the progress.